Change structure of the files; separate different parts into different files
